+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# 21.0.2
+# 21.1.1
+
+- Fix [#8](https://github.com/shinolab/autd3-rs/issues/8): `phase` parameter of `gain::Bessel` and `gain::Focus` have no effect
+
+# 21.1.0
 
 - Fix [#7](https://github.com/shinolab/autd3-rs/issues/7): `on_lost` and `on_err` callback don't work
 - Add `with_err_handler` to `autd3-link-soem::SOEM`